Agente_comunicador/requirements.txt
# Dependencias principales para el agente de voz con LangGraph
# Instalar con: pip install -r requirements.txt

# LangChain y componentes
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-tavily>=0.1.0
langgraph>=0.1.0

# Variables de entorno
python-dotenv>=1.0.0

# Audio y reconocimiento de voz
SpeechRecognition>=3.10.0
gtts>=2.4.0
playsound>=1.3.0
PyAudio>=0.2.11

# APIs y requests
requests>=2.31.0

# Utilidades adicionales (opcionales pero recomendadas)
typing-extensions>=4.8.0
pydantic>=2.0.0

# Para desarrollo (opcional)
jupyter>=1.0.0
ipython>=8.0.0